Rename Tools to be postfixed 'Tool'
