+++ src/lib/components/Tools/DrawTool.svelte
