+++ src/lib/components/Tools/EraseTool.svelte
