+++ src/lib/components/Tools/PanTool.svelte
